bpfman-operator: Add test helper for BpfmanConfigReconciler setup
Introduce a new test helper function `setupTestEnvironment` to
standardise the setup process for BpfmanConfigReconciler tests. This
function handles the initialisation of the fake client, the registration
of necessary types with the runtime scheme, and the configuration of the
reconciler.

The helper also allows for specific OpenShift testing by conditionally
registering the SecurityContextConstraints (SCC) type.

This helper will be used in subsequent commits to refactor existing
tests and to introduce a new test for OpenShift-specific functionality.